requirements.txt
pytest~=8.4.1
httpx~=0.27.0
SQLAlchemy~=2.0.43
PyJWT~=2.10.1
python-dotenv~=1.0.1
pydantic~=2.5.2
uvicorn~=0.29.0
fastapi~=0.111.0
python-jose
bcrypt~=4.3.0
pytest-asyncio>=0.23.0
asyncpg
aiosqlite~=0.20.0
slowapi
sentry-sdk[fastapi]==1.40.6